YarnSpinner intregration V0.22
Loading in file to edit, idk if this was mentioned in last commit
Getting the program at runtime compile yarn files in an outside folder to run dialogues - still working on
Learned more about yarn spinner, will probably start writing and doing tests on commands and such with dialogue when the above issue is resolved.
Lots of scripts need reworking from the yarn spinner example scene to work with this program given its probably entirely UI based.
Need to get that dialoguerunner to properly load in files.